Move to patch object without resourceVersion
This is to work around kubernetes being super aggressive with namespace
deletes.

A namespace delete cascades to all the resources contained within that
namespace, including CRD resources.
Kubernetes retries these deletes on a tight loop which means the
CRD code can't get an optimistic lock on the resourceVersion and can't
remove the finalizer from the object to allow deletion.

Currently running multiple instances of the same controller isn't
supported anyway so this _should_ be fine.